% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_cnst}
\alias{get_cnst}
\alias{RMUMPS_PERM_AMD}
\alias{RMUMPS_PERM_AMF}
\alias{RMUMPS_PERM_SCOTCH}
\alias{RMUMPS_PERM_PORD}
\alias{RMUMPS_PERM_METIS}
\alias{RMUMPS_PERM_QAMD}
\alias{RMUMPS_PERM_AUTO}
\title{get value of a c++ macro constant}
\usage{
get_cnst(s)
}
\arguments{
\item{s}{character scalar, name of the constant}
}
\value{
integer scalar, value of the sonstant
}
\description{
Return an integer number defined as c++ macro if it is present in internal dictionary. The input parameter is the name of the constant to be returned. If the name is not found in the dictionary, an error is thrown. Constants currently present in the dictionary are following: RMUMPS_PERM_AMD, RMUMPS_PERM_AMF, RMUMPS_PERM_SCOTCH, RMUMPS_PERM_PORD, RMUMPS_PERM_METIS, RMUMPS_PERM_QAMD and RMUMPS_PERM_AUTO.
}
\examples{
get_cnst("RMUMPS_PERM_AMD")
}
