% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rmumps__del_ptr}
\alias{Rmumps__del_ptr}
\title{Delete via Pointer}
\usage{
Rmumps__del_ptr(pm)
}
\arguments{
\item{pm}{pointer of type XPtr<Rmumps>, object to be deleted}
}
\description{
This is a C wrapper to \code{Rmumps::~Rmumps()} destructor. Available in R too.
In C++ code can be used as \code{rmumps::Rmumps__del_ptr(pm)}
}
