% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rmumps__get_permutation}
\alias{Rmumps__get_permutation}
\title{Get Permutation Parameter}
\usage{
Rmumps__get_permutation(pm)
}
\arguments{
\item{pm}{pointer of type XPtr<Rmumps>, object having sparse matrix permuted according to some method.}
}
\value{
integer defining permutation method used before matrix decomposition.
}
\description{
This is a C wrapper to \code{Rmumps::get_permutation()} method. Available in R too.
In C++ code can be used as \code{rmumps::Rmumps__get_permutation(pm)}
}
