% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rmumps__set_mat_ptr}
\alias{Rmumps__set_mat_ptr}
\title{Set Matrix via Pointer}
\usage{
Rmumps__set_mat_ptr(pm, pa)
}
\arguments{
\item{pm}{pointer of type XPtr<Rmumps>, object having sparse matrix to be replaced with second parameter}

\item{pa}{pointer of type XPtr<double>, value vector from sparse triplet providing a new matrix. Structure of the new matrix must be identical to the old one. That's why there is no need to provide i and j for the new triplet.}
}
\description{
This is a C wrapper to \code{Rmumps::set_mat_ptr(a)} method. Available in R too.
In C++ code can be used as \code{rmumps::Rmumps__set_mat_ptr(pm)}. Using this method invalidates previous numeric decomposition (but not symbolic one).
}
