% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rmumps__set_permutation}
\alias{Rmumps__set_permutation}
\title{Set Permutation Parameter}
\usage{
Rmumps__set_permutation(pm, permutation)
}
\arguments{
\item{pm}{pointer of type XPtr<Rmumps>, object having sparse matrix permuted according to a chosen method.}

\item{permutation}{integer one of predefined constants (cf. \code{\link{RMUMPS_PERM}}). Setting a new permutation invalidates current symbolic and numeric matrix decompositions.}
}
\description{
This is a C wrapper to \code{Rmumps::set_permutation(permutation)} method. Available in R too.
In C++ code can be used as \code{rmumps::Rmumps__set_permutation(pm, permutation)}
}
