% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rmumps__solveptr}
\alias{Rmumps__solveptr}
\title{Solve via Pointer}
\usage{
Rmumps__solveptr(pobj, pb, lrhs, nrhs)
}
\arguments{
\item{pobj}{pointer of type XPtr<Rmumps>, object having sparse matrix}

\item{pb}{pointer of type XPtr<double>, vector or dense matrix of rhs}

\item{lrhs}{integer, leading dimension in pb}

\item{nrhs}{integer, number of rhs to solve.}
}
\description{
This is a C wrapper to \code{Rmumps::solveptr()} method. Available in R too.
In C++ code can be used as \code{rmumps::Rmumps__solveptr(pobj, pb, lrhs, nrhs)}
}
