% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rmumps__triplet}
\alias{Rmumps__triplet}
\title{Explore via Triplet}
\usage{
Rmumps__triplet(pm)
}
\arguments{
\item{pm}{pointer of type XPtr<Rmumps>, object having sparse matrix to be explored}
}
\value{
a list with sparse triplet described with fields i, j, v
}
\description{
This is a C wrapper to \code{Rmumps::triplet()} method. Available in R too.
In C++ code can be used as \code{rmumps::Rmumps__triplet(pm)}
}
