% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{RMUMPS_PERM}
\alias{RMUMPS_PERM}
\alias{RMUMPS_PERM_AMD}
\alias{RMUMPS_PERM_AMF}
\alias{RMUMPS_PERM_SCOTCH}
\alias{RMUMPS_PERM_PORD}
\alias{RMUMPS_PERM_METIS}
\alias{RMUMPS_PERM_QAMD}
\alias{RMUMPS_PERM_AUTO}
\title{Exported Constants}
\description{
Integer constants defining permutation types and exported from rmumps are following:
\itemize{
\item{\code{RMUMPS_PERM_AMD}}
\item{\code{RMUMPS_PERM_AMF}}
\item{\code{RMUMPS_PERM_SCOTCH}}
\item{\code{RMUMPS_PERM_PORD}}
\item{\code{RMUMPS_PERM_METIS}}
\item{\code{RMUMPS_PERM_QAMD}}
\item{\code{RMUMPS_PERM_AUTO}}
}
They are all regrouped in a named vector \code{RMUMPS_PERM} where names are items above and values are corresponding constants.
}
\examples{
am=rmumps::Rmumps$new(slam::as.simple_triplet_matrix(diag(1:3)))
am$set_permutation(RMUMPS_PERM_SCOTCH)
am$solve(1:3)
}
